introduce package-info.js to wrap details from the app's package.json
