+++ app/package-info.d.ts
@@ -1,0 +1,4 @@
+export function getProductName(): string
+export function getCompanyName(): string
+export function getVersion(): string
+export function getBundleID(): string
