+++ script/package.ts
@@ -4,13 +4,12 @@ import * as fs from 'fs-extra'
 import * as cp from 'child_process'
 import * as path from 'path'
 import * as electronInstaller from 'electron-winstaller'
+import { getProductName, getCompanyName } from '../app/package-info'
 import {
   getDistRoot,
   getDistPath,
-  getProductName,
   getOSXZipPath,
   getWindowsIdentifierName,
-  getCompanyName,
   getWindowsStandaloneName,
   getWindowsInstallerName,
   shouldMakeDelta,
