+++ script/test-setup.ts
@@ -3,7 +3,8 @@
 import * as fs from 'fs'
 import * as cp from 'child_process'
 import { getLogFiles } from './review-logs'
-import { getProductName, getDistPath } from './dist-info'
+import { getProductName } from '../app/package-info'
+import { getDistPath } from './dist-info'
 
 const isFork = process.env.CIRCLE_PR_USERNAME
